requirements.txt
asgiref==3.2.10
beautifulsoup4==4.9.1
decorator==4.4.2
Django==3.0.8
django-autoslug==1.9.7
django-bootstrap4==2.2.0
django-ckeditor==5.9.0
django-email-verification==0.0.6
django-js-asset==1.2.2
Pillow==7.2.0
pytz==2020.1
self==0.0.0
soupsieve==2.0.1
sqlparse==0.3.1
